+++ modules/angular2/test/transform/directive_processor/parameter_metadata/soup.dart
@@ -1,0 +1,8 @@
+library dinner.soup;
+
+import 'package:angular2/src/core/annotations/annotations.dart';
+
+@Component(selector: '[soup]')
+class SoupComponent {
+  SoupComponent(@Tasty String description, @Inject(Salt) salt);
+}
