+++ modules/angular2/test/transform/integration/basic_bind_files/bar.dart
