+++ modules/angular2/test/transform/integration/basic_bind_files/expected/bar.ngDeps.dart
