+++ modules/angular2/test/transform/integration/chained_deps_files/expected/bar.ngDeps.dart
