+++ modules/angular2/test/transform/integration/chained_deps_files/expected/foo.ngDeps.dart
