+++ modules/angular2/test/transform/integration/chained_deps_files/foo.dart
