+++ modules/angular2/test/transform/integration/list_of_types_files/bar.dart
