+++ modules/angular2/test/transform/integration/simple_annotation_files/expected/bar.ngDeps.dart
