+++ modules/angular2/test/transform/integration/simple_annotation_files/expected/index.dart
