+++ modules/angular2/test/transform/integration/simple_annotation_files/index.dart
