+++ modules/angular2/test/transform/integration/synthetic_ctor_files/bar.dart
