+++ modules/angular2/test/transform/integration/synthetic_ctor_files/expected/bar.ngDeps.dart
