+++ modules/angular2/test/transform/integration/two_annotations_files/bar.dart
