+++ modules/angular2/test/transform/integration/two_deps_files/bar.dart
