+++ modules/angular2/test/transform/integration/two_deps_files/expected/bar.ngDeps.dart
