+++ modules/angular2/test/transform/reflection_remover/all_tests.dart
@@ -1,0 +1,21 @@
+library angular2.test.transform.reflection_remover;
+
+import 'package:analyzer/analyzer.dart';
+import 'package:angular2/src/transform/reflection_remover/codegen.dart';
+import 'package:angular2/src/transform/reflection_remover/rewriter.dart';
+import 'package:unittest/unittest.dart';
+
+import 'reflection_remover_files/expected/index.dart' as expected;
+import '../common/read_file.dart';
+
+void allTests() {
+  var codegen = new Codegen('web/index.dart', 'web/index.ngDeps.dart');
+
+  test('Remover', () {
+    var code =
+        readFile('reflection_remover/reflection_remover_files/index.dart');
+    var output =
+        new Rewriter(code, codegen).rewrite(parseCompilationUnit(code));
+    expect(output, equals(expected.code));
+  });
+}
