+++ modules/angular2/test/transform/reflection_remover/reflection_remover_files/expected/index.dart
